Move wrapper into mapper package
