+++ sadu-mapper/src/main/java/de/chojo/sadu/mapper/DefaultMapper.java
@@ -10,7 +10,7 @@ import de.chojo.sadu.core.exceptions.ThrowingBiFunction;
 import de.chojo.sadu.core.types.SqlType;
 import de.chojo.sadu.mapper.rowmapper.RowMapper;
 import de.chojo.sadu.mapper.util.Results;
-import de.chojo.sadu.wrapper.util.Row;
+import de.chojo.sadu.mapper.wrapper.Row;
 
 import java.math.BigDecimal;
 import java.sql.ResultSetMetaData;
