+++ sadu-queries/src/main/java/de/chojo/sadu/queries/stages/execution/reading/ReaderImpl.java
@@ -20,7 +20,7 @@ import de.chojo.sadu.queries.query.TokenizedQuery;
 import de.chojo.sadu.queries.stages.execution.writing.CalledSingletonQueryImpl;
 import de.chojo.sadu.queries.stages.results.reading.MultiResult;
 import de.chojo.sadu.queries.stages.results.reading.SingleResult;
-import de.chojo.sadu.wrapper.util.Row;
+import de.chojo.sadu.mapper.wrapper.Row;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
